# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is a sample server AutomartAPI. Showing the AutoMartAPI backend APIS",
    "version": "1.0.0",
    "title": "AutoMart-APi",
    "termsOfService": "localhost:3000",
    "contact": {
      "email": "williamtrans2011@gmail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "localhost:3000",
  "basePath": "/api/v1",
...
  "schemes": [
    "https",
    "http"
  ],
  "paths": {
...
    "/car/{carId}": {
      "get": {
        "tags": [
          "car"
        ],
        "summary": "Find car by ID",
        "description": "Returns a single car",
        "operationId": "getCarById",
        "produces": [
          "application/xml",
          "application/json"
        ],
        "parameters": [
          {
            "name": "carId",
            "in": "path",
            "description": "ID of car to return",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Car"
            }
          },
          "404": {
            "description": "Car not found"
          }
        }
      },
      "delete": {
        "tags": [
          "car"
        ],
        "summary": "Deletes a car",
        "description": "",
        "operationId": "deleteCar",
        "produces": [
          "application/xml",
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string"
          },
          {
            "name": "carId",
            "in": "path",
            "description": "Car id to delete",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "404": {
            "description": "Car not found"
          },
          "200": {
            "description": "Car deleted successfully"
          }
        }
      }
    },
...
  },
  "securityDefinitions": {
    "petstore_auth": {
      "type": "oauth2",
      "authorizationUrl": "http://petstore.swagger.io/oauth/dialog",
      "flow": "implicit",
      "scopes": {
        "write:cars": "modify cars in your account",
        "read:cars": "read your cars"
      }
    },
    "api_key": {
      "type": "apiKey",
      "name": "authorization",
      "in": "header"
    }
  },
  "definitions": {
...
    "Car": {
      "type": "object",
      "required": [
        "model",
        "manufacturer",
        "price",
        "body_type",
        "state"
      ],
      "properties": {
        "model": {
          "type": "string",
          "example": "x-6"
        },
        "manufacturer": {
          "type": "string",
          "example": "benz"
        },
        "price": {
          "type": "number"
        },
        "body_type": {
          "type": "string",
          "example": "[car, suv, pick-up]"
        },
        "state": {
          "type": "string",
          "description": "car status",
          "enum": [
            "new",
            "used"
          ]
        },
        "status": {
          "type": "string",
          "default": "available",
          "description": "pet status in the store",
          "enum": [
            "available",
            "pending",
            "sold"
          ]
        }
      }
    },
...
  }
}